refactor: Comprehensive architecture refactoring
## Test Organization (One Class Per File)
- Split test_models.py into 5 separate files (test_search_result.py, test_api_search_result.py, etc.)
- Split test_content_scraper.py into 3 files by scenario (test_success_cases.py, test_error_cases.py, test_edge_cases.py)
- Split test_duckduckgo_client.py into 2 files (with/without library)
- Split test_search_service.py into 2 files (with/without Serper key)
- Result: 44 tests passing, improved test organization and maintainability

## Module Reorganization
- Split api_tools.py (405 lines) into focused modules:
  - models/responses/api_responses.py (API response models)
  - services/search_orchestrator.py (search logic with fallbacks)
  - tools/api_tools_setup.py (MCP tool setup)
- Split health.py (281 lines) into focused modules:
  - models/responses/health_responses.py (health response formatters)
  - endpoints/health_endpoints.py (endpoint handlers)
  - endpoints/demo_client.py (demo client serving)
  - endpoints/app_factory.py (FastAPI app factories)
- Reorganized models/ directory:
  - models/domain/ (SearchResult, APISearchResult)
  - models/responses/ (API responses, health responses)
- All original files now import from new locations with deprecation warnings for backward compatibility

## Consolidation
- Removed demo_server.py (kept simple_demo.py as single demo server)
- Removed debug_search.py (debug script)
- Removed old_tests/ directory (5 obsolete test files)

## Benefits
- Single responsibility principle enforced throughout
- Test files follow "one class per file" standard
- Clear separation: domain models vs API responses vs endpoints vs services
- Reduced largest files from 405 and 281 lines to focused <150 line modules
- Better discoverability and maintainability
- All existing imports still work via backward compatibility shims
