+++ docker/models/domain/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) 2024 Travis Frisinger
+#
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+
+"""Domain models."""
