+++ docker/tests/unit/services/content_scraper/test_success_cases.py
@@ -1,0 +1,55 @@
+# Copyright (c) 2024 Travis Frisinger
+#
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+
+"""Unit tests for successful content scraping scenarios."""
+
+from unittest.mock import patch
+
+from services.content_scraper import scrape_search_result
+from tests.builders.search_result_builder import a_search_result
+from tests.factories.http_response_factory import HttpResponseFactory
+
+
+class TestContentScraperSuccess:
+    """Tests for successful content scraping scenarios."""
+
+    @patch("services.content_scraper.requests.get")
+    def test_converts_html_to_markdown_with_title(self, mock_get):
+        # Arrange
+        result = a_search_result().with_title("Test Article").build()
+        mock_get.return_value = HttpResponseFactory.html_with_title("Test Article")
+
+        # Act
+        scraped = scrape_search_result(result)
+
+        # Assert
+        assert scraped.content.startswith("# Test Article")
+        assert "*Source: https://example.com/test*" in scraped.content
+
+    @patch("services.content_scraper.requests.get")
+    def test_wraps_plaintext_in_code_blocks(self, mock_get):
+        # Arrange
+        result = a_search_result().build()
+        mock_get.return_value = HttpResponseFactory.plaintext("Plain text content")
+
+        # Act
+        scraped = scrape_search_result(result)
+
+        # Assert
+        assert "```" in scraped.content
+        assert "Plain text content" in scraped.content
+
+    @patch("services.content_scraper.requests.get")
+    def test_handles_pdf_files_with_message(self, mock_get):
+        # Arrange
+        result = a_search_result().build()
+        mock_get.return_value = HttpResponseFactory.pdf()
+
+        # Act
+        scraped = scrape_search_result(result)
+
+        # Assert
+        assert "binary file" in scraped.content
+        assert "application/pdf" in scraped.content
